Fixed workflow triggers (#1)
* Fix workflow file and various linting issues

Fix the workflow to build using the Ubuntu 20.04 and 22.04 with the
stable, beta and nightly toolchains for Rust. This ensures that we are
keeping up with Rust changes. Also the wrap_comments and comment_width
parameters in rustfmt need to be run on the nightly toolchain since they
are not considered stable yet.

Also fixed the char conversion. Since one some platforms a char is uint8
and on others it is int8.

* Simplify FFI ownership

Instead of converting to and from raw pointers using Box::into_raw and
Box::from_raw, just keep the Box around and use the built-in Drop trait.
Rust will see this as dead code but we can suppress the error using the
allow() macro.