Use bindgen to generate the FFI bindings
Fixes #3

Some implementations of `secapi` break the ABI of the reference
implementation which causes nasty undefined behavior when using the
checked-in bindings based off the reference implementation. Instead of
checking the bindings into the repo, let's opt for generating them at
build-time with bindgen. This way, we can be confident that the bindings
are ABI compatible with any linked implementation of `secapi`.