Use bindgen to generate the FFI bindings (#23)
* Some implementations of `secapi` break the ABI of the reference
implementation which causes nasty undefined behavior when using the
checked-in bindings based off the reference implementation. Instead of
checking the bindings into the repo, let's opt for generating them at
build-time with bindgen. This way, we can be confident that the bindings
are ABI compatible with any linked implementation of `secapi`.

* Use c_char type for Rights::id
This isn't exactly a great type to choose here. This should probably be
a String or CString. For the sake of compatibility with `saclient` and
because it isn't really used for anything, we will just leave as
`c_char`.